refactor: UserContext를 lib/user-context로 이동 (#82)
